requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
async-timeout==4.0.3
    # via redis
certifi==2024.6.2
    # via requests
charset-normalizer==3.3.2
    # via requests
idna==3.7
    # via requests
joblib==1.4.2
    # via scikit-learn
numpy==2.0.0
    # via
    #   pandas
    #   scikit-learn
    #   scipy
pandas==2.2.2
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via -r requirements.in
redis==5.0.7
    # via spotipy
requests==2.32.3
    # via spotipy
scikit-learn==1.5.0
    # via -r requirements.in
scipy==1.14.0
    # via
    #   -r requirements.in
    #   scikit-learn
six==1.16.0
    # via python-dateutil
spotipy==2.24.0
    # via -r requirements.in
threadpoolctl==3.5.0
    # via scikit-learn
tzdata==2024.1
    # via pandas
urllib3==2.2.2
    # via
    #   requests
    #   spotipy
